requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-heroku==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djongo==1.3.6
dnspython==2.1.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.1
mysqlclient==2.0.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
py3-validate-email==1.0.0
PyJWT==1.7.1
pymongo==3.11.4
python-decouple==3.4
pytz==2021.1
sqlparse==0.2.4
whitenoise==5.2.0
